refactor: relocate evals under project
