+++ src/phoenix/server/api/routers/span_handler.py
@@ -1,7 +1,7 @@
 import asyncio
 import gzip
 from functools import partial
-from typing import AsyncIterator, Optional
+from typing import AsyncIterator
 
 import opentelemetry.proto.trace.v1.trace_pb2 as otlp
 from starlette.endpoints import HTTPEndpoint
@@ -9,7 +9,7 @@ from starlette.requests import Request
 from starlette.responses import Response, StreamingResponse
 from starlette.status import HTTP_404_NOT_FOUND, HTTP_422_UNPROCESSABLE_ENTITY
 
-from phoenix.core.evals import Evals
+from phoenix.core.project import DEFAULT_PROJECT_NAME
 from phoenix.core.traces import Traces
 from phoenix.server.api.routers.utils import df_to_bytes, from_iso_format
 from phoenix.trace.dsl import SpanQuery
@@ -21,7 +21,6 @@ from phoenix.utilities import query_spans
 
 class SpanHandler(HTTPEndpoint):
     traces: Traces
-    evals: Optional[Evals] = None
 
     async def post(self, request: Request) -> Response:
         try:
@@ -45,20 +44,22 @@ class SpanHandler(HTTPEndpoint):
     async def get(self, request: Request) -> Response:
         payload = await request.json()
         queries = payload.pop("queries", [])
+        if not (project := self.traces.get_project(DEFAULT_PROJECT_NAME)):
+            return Response(status_code=HTTP_404_NOT_FOUND)
         loop = asyncio.get_running_loop()
         valid_eval_names = (
             await loop.run_in_executor(
                 None,
-                self.evals.get_span_evaluation_names,
+                project.get_span_evaluation_names,
             )
-            if self.evals
+            if project
             else ()
         )
         try:
             span_queries = [
                 SpanQuery.from_dict(
                     query,
-                    evals=self.evals,
+                    evals=project,
                     valid_eval_names=valid_eval_names,
                 )
                 for query in queries
