+++ src/phoenix/server/api/types/Span.py
@@ -10,6 +10,7 @@ from strawberry import ID, UNSET
 from strawberry.types import Info
 
 import phoenix.trace.schemas as trace_schema
+from phoenix.core.project import DEFAULT_PROJECT_NAME
 from phoenix.metrics.retrieval_metrics import RetrievalMetrics
 from phoenix.server.api.context import Context
 from phoenix.server.api.types.DocumentRetrievalMetrics import DocumentRetrievalMetrics
@@ -144,12 +145,14 @@ class Span:
         self,
         info: Info[Context, None],
     ) -> List[SpanEvaluation]:
-        if not (evals := info.context.evals):
+        if not (traces := info.context.traces):
+            return []
+        if not (project := traces.get_project(DEFAULT_PROJECT_NAME)):
             return []
         span_id = SpanID(str(self.context.span_id))
         return [
             SpanEvaluation.from_pb_evaluation(evaluation)
-            for evaluation in evals.get_evaluations_by_span_id(span_id)
+            for evaluation in project.get_evaluations_by_span_id(span_id)
         ]
 
     @strawberry.field(
@@ -164,12 +167,14 @@ class Span:
         self,
         info: Info[Context, None],
     ) -> List[DocumentEvaluation]:
-        if not (evals := info.context.evals):
+        if not (traces := info.context.traces):
+            return []
+        if not (project := traces.get_project(DEFAULT_PROJECT_NAME)):
             return []
         span_id = SpanID(str(self.context.span_id))
         return [
             DocumentEvaluation.from_pb_evaluation(evaluation)
-            for evaluation in evals.get_document_evaluations_by_span_id(span_id)
+            for evaluation in project.get_document_evaluations_by_span_id(span_id)
         ]
 
     @strawberry.field(
@@ -180,10 +185,12 @@ class Span:
         info: Info[Context, None],
         evaluation_name: Optional[str] = UNSET,
     ) -> List[DocumentRetrievalMetrics]:
-        if not self.num_documents or not (evals := info.context.evals):
+        if not self.num_documents or not (traces := info.context.traces):
+            return []
+        if not (project := traces.get_project(DEFAULT_PROJECT_NAME)):
             return []
         span_id = SpanID(str(self.context.span_id))
-        all_document_evaluation_names = evals.get_document_evaluation_names(span_id)
+        all_document_evaluation_names = project.get_document_evaluation_names(span_id)
         if not all_document_evaluation_names:
             return []
         if evaluation_name is UNSET:
@@ -194,7 +201,7 @@ class Span:
             evaluation_names = [evaluation_name]
         retrieval_metrics = []
         for name in evaluation_names:
-            evaluation_scores = evals.get_document_evaluation_scores(
+            evaluation_scores = project.get_document_evaluation_scores(
                 span_id=span_id,
                 evaluation_name=name,
                 num_documents=self.num_documents,
